pyproject.toml
[tool.poetry]
name = "baidu-ultr"
version = "0.1.0"
description = ""
authors = ["Philipp Hager <philipp.konstantin.hager@gmail.com>"]
readme = "README.md"
packages = [{ include = "baidu-ultr" }]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.0"
transformers = { extras = ["torch"], version = "^4.34.0" }
scikit-learn = "^1.3.1"
wget = "^3.2"
hydra-core = "^1.3.2"
safetensors = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
